Refactor io.higherkindness to higherkindness
